runtime: Use pv_runtime_take_from_host() for ld.so
In a mutable sysroot, overwrite ld.so with a symlink into /run/host.
We overwrite both the interoperable path (to avoid long chains of
symlinks in the common case) and the canonicalized path (to make sure
nothing can end up finding the container's ld.so, which is incompatible
with the libc we have dropped in).

If building a long bwrap command-line, we have to mount over the
canonicalized path, and just live with the fact that there will be
a long chain of symlinks leading to it.

Signed-off-by: Simon McVittie <smcv@collabora.com>