wrap: Add an --only-prepare mode
This is primarily useful for testing and development: it lets us
generate the temporary sysroot and inspect its contents, even if we
are in a container (such as Docker or Flatpak) that does not give us
the necessary privileges to run bubblewrap and create a new user
namespace.

In particular, this will make it more straightforward for us to check
some of our assumptions when running on a non-Debian-derived host
system.

Signed-off-by: Simon McVittie <smcv@collabora.com>