runtime: Move pressure-vessel overrides into /usr
Flatpak is not going to give us control over the root directory when
using sub-sandboxing - the most we're likely to get is control over
which directory is chosen to be /usr. Prepare for this by using
/usr/lib/pressure-vessel/overrides instead of /overrides for the things
we overwrite.

To be nice to developers and steam-runtime-tools, put a symlink
overrides -> usr/lib/pressure-vessel/overrides at the root, as a hint
for where to look.

Signed-off-by: Simon McVittie <smcv@collabora.com>