runtime: Break hard-link for lock file when copying runtime
Because we do a "cheap copy" with hard-links where possible, .ref is a
hard-link to its counterpart in the original (non-copied) runtime, so a
lock on one is a lock on the other. We hold a lock on the original
runtime while we do setup, so we would never actually delete the
temporary runtimes unless we break the hard-link by re-creating the
lock file.

Resolves: T22035
Signed-off-by: Simon McVittie <smcv@collabora.com>