tests: Rename pressure-vessel log file in artifacts
This log file contains pressure-vessel's debug messages (which happen
outside the scout container), not just the output of inside-scout.py,
so calling it inside-scout.log is misleading.

If we're running pressure-vessel-wrap --only-prepare, then we never
actually get as far as running inside-scout.py or entering the container,
so the log file name is doubly misleading.

Signed-off-by: Simon McVittie <smcv@collabora.com>