wrap: Import Steam compat tool data
This is necessary for Steam "compat tools" such as Proton.

STEAM_COMPAT_DATA_PATH points to data associated with the current
(compat tool, game) pair, for example the ${WINEPREFIX}.

STEAM_COMPAT_TOOL_PATH doesn't exist yet, but I'm assuming that it will
be set by some later version of the Steam client to point to Proton
itself.

STEAM_COMPAT_CLIENT_INSTALL_PATH points to the Steam installation. This
is hopefully in the home directory, so hopefully we bind-mount it anyway;
but if we don't, we should.

Signed-off-by: Simon McVittie <smcv@collabora.com>