build: Add SANITIZE_CFLAGS build variable
When building with things like AddressSanitizer, we want to apply
instrumentation to the libcapsule library itself, the libcapsule
tools, and the unit tests. However, we do not want to instrument the
tiny libraries that are used as test fixtures, because giving them a
dependency on libasan.so.5 would confuse libcapsule when testing them.

Signed-off-by: Simon McVittie <smcv@collabora.com>