utils: Allow selecting library comparators from a string
This is a piece of necessary infrastructure for exposing this as a
command-line option, or even as per-library metadata. We were originally
going to do the equivalent of "versions,name,symbols" unconditionally,
but it looks as though that could break more than it fixes, so let's
be a bit more cautious.

Signed-off-by: Simon McVittie <smcv@collabora.com>