capture-libs: Make library comparison configurable on the command-line
This is primarily useful for testing and experimenting. Using versions
by default in preference to filenames (--compare-by="versions,name")
looks like it might be viable, but is a destabilising change that we
should test more before considering a change of defaults.

Meanwhile, counting symbols as a fallback (--compare-by="...,symbols")
does not look as safe as we had hoped, because if a library maintainer
has cleaned up their ABI by hiding private symbols without adding any
new symbols, we will sort libraries in exactly the wrong order - and
in reality, that seems to be what has happened in several libraries,
for example libX11.so.6 and libXfixes.so.3.

Signed-off-by: Simon McVittie <smcv@collabora.com>