capture-libs: Add --library-knowledge option
This gives capsule-capture-libs a source of library-specific knowledge.
For example, if we know that:

  * libgcc_s.so.1 is installed with an unhelpful name, but it uses
    versioned symbols the way you'd hope;
  * libdbus-1.so.3 is installed with a helpful libtool-style name,
    but has private symbols that defeat our current simplistic
    comparisons

then we can express that as:

    [Library libgcc_s.so.1]
    CompareBy=versions;symbols;

    [Library libdbus-1.so.3]
    CompareBy=name;

A runtime that contains a known set of libraries would be a good place
to put library-specific knowledge about those libraries. For example,
the maintainers of the Steam Runtime know what libraries it contains,
and are well-placed to compare those libraries with their counterparts
in mainstream distributions.

Signed-off-by: Simon McVittie <smcv@collabora.com>