run-in-sysroot: Copy host /etc/passwd, /etc/group
Otherwise, any test that refers to the username or home directory will
fail with a warning. Prior to GLib 2.35.9 (in particular, in GLib 2.32)
this happens in g_get_any_init_do(), which is also called to get the
TMPDIR, so any test that wants to create a temporary directory will
fail like this.

Signed-off-by: Simon McVittie <smcv@collabora.com>