tests: Run s-r-s-i on host with LD_BIND_NOW=1
If we allow lazy binding, we can succeed in loading a GIO module from
the host (which depends on a newer GObject than the one bundled with
pressure-vessel), then find that a symbol reference fails at the time
that we try to actually use it. Make it fail early, instead.

In particular, this is helpful on SteamOS 2 'brewmaster'.

Signed-off-by: Simon McVittie <smcv@collabora.com>