tests: Ignore inability to load libOSMesa.so.8
This is included in prerelease versions of SteamRT 2 'soldier',
but because it uses the post-gcc-5 C++ ABI of libLLVM-7.so.1, it's
incompatible with versions of libLLVM-7.so.1 that were compiled to be
compatible with the gcc 4 C++ ABI, and in particular the one in
SteamOS 2 'brewmaster'.

This and the previous commit are sufficient to make tests/containers.py
pass on SteamOS 2 'brewmaster', if we also configure libxcb.so.1 to
be chosen by counting symbols (T22538) and libgcc_s.so.1 to be chosen
by counting verdefs (T16561).

Signed-off-by: Simon McVittie <smcv@collabora.com>