populate-depot: If SDK is requested, include SDK runtime and debug symbols
The SDK runtime tarball has almost the same content as the sysroot,
but arranged in the same way as the Platform, and with some compat
symlinks that make it more convenient to use on non-Debian-derived
host systems (in particular Arch Linux, in which some libraries and
executables have a non-standard ld.so path).

The detached debug symbols are necessary if you want to be able to run
gdb inside the pressure-vessel container.

Signed-off-by: Simon McVittie <smcv@collabora.com>