Add flatpak-context, flatpak-exports from Flatpak
A lot of the code added here is #if 0. Normally we don't commit
commented-out or #ifdef'd out code, but in this case it helps tools
like gvimdiff to have enough context to figure out which parts of the
file in Flatpak correspond to which parts of the file in
pressure-vessel.

Signed-off-by: Simon McVittie <smcv@collabora.com>