utils: Add public/private knowledge about symbol versions and symbols
We are now able to list the symbol versions and/or symbols that are
known to be public/private.

This is especially useful for comparing libraries that removed symbol
versions, or symbols, from one release to another. As this already
happened multiple times in the past, like for example with
`libdrm_nouveau.so.2` or `libedit.so.2`.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>