Initial support for custom graphics provider path and Flatpak
pressure-vessel had a few hard coded assumptions about the host system
and where to look at for the graphics provider.

For example it was assumed that the root was the same host system used
by both pressure-vessel current environment (where pressure-vessel-wrap
runs) and also later on by `bwrap`.

But when we run pressure-vessel in a Flatpak container, the final
`bwrap` command is expected to be executed against the real host system
(mounted in `/run/host`, in Flatpak).

Also when we are in a Flatpak container we might want to pick the
graphics stack from the host system (`/run/host`), from the Flatpak
container itself or it could even be a completely separate one.
(T22373, T22371)

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>