wrap: Don't populate environment of FlatpakBwrap that will be merged
The default for a FlatpakBwrap object is to copy the calling
environment as a base for the environment that it will build up.
However, that's wrong if we are going to merge more than one
FlatpakBwrap together: the merged FlatpakBwrap's argv will be appended
to the destination FlatpakBwrap, but the merged envp will overwrite the
corresponding variables in the destination.

Avoid this by making sure that every time we merge two FlatpakBwrap
objects, one of them has an empty environment.

Resolves: T23422
Signed-off-by: Simon McVittie <smcv@collabora.com>