bwrap: Build arguments to bind /usr if already in a container
The filesystem that is providing /usr might not be mounted
in the same place in the current namespace, and on the host
system where bwrap will be run. If it isn't, we need
to use one path to do file I/O when inspecting it, but pass
a different path to bwrap.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>