utils: improve path conversion from current namespace to host
When the current namespace is a Flatpak environment launched with the
`--filesystem` option to mount a directory that is under the user's
home, e.g. `--filesystem=~/Games`, we don't want to convert the path to
`~/.var/app/com.valvesoftware.Steam/[...]`, because on the host we
expect to have the same path as we had in the Flatpak environment.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>