wrap: Use flatpak-spawn to launch bwrap if we are in a Flatpak container
When pressure-vessel is launched from inside a Flatpak container we are
not allowed to create a parallel bwrap container.

Using flatpak-spawn --host we can workaround this limitation.

Please note that the permission to talk to "org.freedesktop.Flatpak" is
required for this to work. For example you can use:
`flatpak run --talk-name=org.freedesktop.Flatpak
com.valvesoftware.Steam`

Hopefully in the future we will be able to avoid the permission
requirement of "org.freedesktop.Flatpak".

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>