wrap, adverb: Move "wrap with xterm" code into the adverb
Looking at environment variables is still part of pv-wrap, but the
command-line part now happens in pv-adverb. This means the launching
scripts for the "session" mode, which use pv-launch to run the main
game command in a previously-set-up pv-launcher environment, can
direct pv-adverb to run an interactive shell instead of or in addition
to the actual game.

Signed-off-by: Simon McVittie <smcv@collabora.com>