runtime: Load basic NSS plugins along with host system glibc
The basic Name Service Switch plugins that ship with glibc are entitled
to assume that they are used alongside a corresponding glibc, and can
use private symbols from it.

This is not a complete set, only the ones required by the nsswitch.conf
that we ship in practice. We can expand coverage later if desired.

For a practical example of this, use soldier (glibc 2.28) on Arch Linux
(glibc 2.32), and run `LD_DEBUG=files id`. soldier's libnss_files.so.2
fails to load with an undefined reference to __libc_readline_unlocked.

This might also resolve various previously-reported bugs with DNS not
working in the scout container:

* https://github.com/ValveSoftware/steam-runtime/issues/226
  (scout on Debian 10)
* https://github.com/ValveSoftware/steam-runtime/issues/264,
  https://github.com/ValveSoftware/steam-runtime/issues/258,
  https://github.com/ValveSoftware/steam-runtime/issues/192
  (scout on Fedora 31)
* https://github.com/ValveSoftware/steam-runtime/issues/262
  (scout on Arch Linux)
* https://github.com/ValveSoftware/steam-runtime/issues/227
  (scout on unspecified distribution)

Signed-off-by: Simon McVittie <smcv@collabora.com>