cheap-copy test: Don't assert that copies preserve ctime
glnx_file_copy_at() preserves atime and mtime, but makes no guarantee
about ctime. In practice this test usually passed anyway, because after
the first file is created, the copy is usually created within the same
second, but this is not guaranteed: if we create the first file just
before a clock tick and copy it just after, the test will fail.

Signed-off-by: Simon McVittie <smcv@collabora.com>