debian: Delete files that were generated by gtk-doc during dist
If we `make dist` with gtk-doc-tools 1.28 from Debian 10 'buster' (the
current stable release), it generates a libcapsule-docs.xml in which
generation of tree_index.sgml is mandatory. This is incompatible with
building this non-GObject library with gtk-doc-tools >= 1.30, in which
tree_index.sgml is only generated if the library contains at least one
GObject type.

Most projects treat the libcapsule-docs.xml generated by gtk-doc as
a template and commit it to git, but in this project we don't
particularly want to maintain it and would prefer to keep regenerating
it during build.

Signed-off-by: Simon McVittie <smcv@collabora.com>