_start-container-in-background: Break as soon as we get the socket
When inside a Flatpak container the socket fifo FD might be opened
multiple times, for example from `flatpak-spawn --host` and from
`pressure-vessel-launcher`.

Normally we expected `pressure-vessel-launcher` to open the FD, write
the `socket=...` into the fifo file and then close the FD.
But if there is at least one other process that keeps the FD open,
`_start-container-in-background` will never finish waiting to read new
lines in the fifo file.

Because of this we just break the loop as soon as we received what we
were waiting for.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>