Add an environment variables locking system
If we want to pass all the possible variables that a user might have set
in its environment, we need a mechanism to lock the variables that
pressure-vessel is editing (e.g. LIBGL_DRIVERS_PATH), variables that we
want to keep unset because will be wrong in the new container (e.g.
FLATPAK_ID) and variables that should inherit their value from the host
system (e.g. DISPLAY).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>