bwrap: Don't bind-mount files into /etc as part of binding /usr
If we want to deal with strange operating systems that use a non-standard
filename for ld.so.cache, we'll have to special-case it. There's
little extra cost to doing this because we already need to understand
how ld.so.cache works, to be able to generate our own with different
search paths.

In situations where we just want a quick container to be able to inspect
the runtime, we can deal with this by mounting all of /etc read-only.

For the final container, we already iterate over all the files in the
runtime's /etc, which will "naturally" include ld.so.cache and
alternatives.

Signed-off-by: Simon McVittie <smcv@collabora.com>