capture-libs: Skip libs with unmet dependencies when "if exists" flag is set
If we have the CAPTURE_FLAG_IF_EXISTS option flag, we should not
propagate an error if we are not able to find some dependencies of a
library. Instead we should just skip it.

This fixes an error that some users reported because they had some
leftovers unused libraries in their system and libcapsule reported an
error while trying to capture their (missing) dependencies.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>