input-device: Add device-type flags
These are what we really want for SDL and Wine: they're broadly
equivalent to the udev ID_INPUT_FOO properties, and identify which
devices are desired at a high level.

Also add test coverage for the device-type-guessing heuristic.
Thanks to various people in Collabora, Codeweavers and #debian-uk for
providing `sudo evemu-describe` output for their devices.

Signed-off-by: Simon McVittie <smcv@collabora.com>