Start using the new JSON-GLib functions "*_with_default"
JSON-Glib 1.6.0 introduced new functions to get members from a JSON
object with a default fallback value.
This allows us to avoid manually checking the existence of a member.

The functions that we currently need, have been backported to allow the
execution on systems with a JSON-GLIB version older than the 1.6.0

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>