Include host's fonts in the newly built container
Soldier and Scout include a limited set of fonts that doesn't cover all
use cases. For example it is missing the necessary glyphs for Asian
languages.

For this reason, instead of just using these fonts, we bind the host's
ones.

This requires an additional configuration file that tells to fontconfig
the location of these fonts.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>