pressure-vessel: Make STEAM_EXTRA_COMPAT_TOOLS_PATHS available
As requested and later documented in
<https://github.com/ValveSoftware/steam-for-linux/issues/6310>,
Steam looks for compatibility tools in locations determined by this
environment variable. Make them available in the container.

When running in Flatpak, this will have no effect for compat tools in
/app, but we should eventually get the Flatpak app's /app by a
different route.

Partial fix for https://github.com/ValveSoftware/steam-runtime/issues/302
(we also need to deal with compat tools in /usr/local/share and
/usr/share, but that can't easily be done here, and is better addressed
when we build the runtimes).

Signed-off-by: Simon McVittie <smcv@collabora.com>