pressure-vessel: Fix sense of assertion, and realign with other checks
This function won't work in Flatpak, so we make sure that we only call
it when we are *not* in Flatpak. The assertion should match that.

This fixes the ability to run:

    ./pressure-vessel/bin/pressure-vessel-wrap -- xterm

when not in a Flatpak environment, which in principle is meant to work
(at least on reasonably normal FHS operating systems), although it isn't
regularly tested.

Thanks to NixOS contributor moben on Github, who spotted this while
debugging.

Signed-off-by: Simon McVittie <smcv@collabora.com>