pressure-vessel: Make sure ~/.steam is available in the container
Previously, we made ~/.steam available in the container as part of
$HOME, but if it was a symlink outside $HOME or contained symlinks
outside $HOME, their targets were not made available in the container.

When using the experimental feature of a per-game home directory,
we correctly made the targets of symlinks inside ~/.steam available,
but not the target of ~/.steam itself (if it was a symlink).

Fixes: https://github.com/ValveSoftware/steam-runtime/issues/291
Signed-off-by: Simon McVittie <smcv@collabora.com>