_v2-entry-point: Optionally shut down and relaunch container
The waitforexitandrun verb is the first opportunity we get to know what
our final environment variables are. It's also the first time we get
run under "adverb" wrappers, if any, such as taskset.

Proton uses a single set of Wine services to run the setup commands,
then shuts them down and relaunches everything for the actual game;
we can do similarly here. However, the startup time cost of doing this
could be significant, so it's currently off by default so that we can
experiment with it. Run with PRESSURE_VESSEL_RELAUNCH_CONTAINER=1 to
try it.

Partially addresses steam-runtime#304.

Resolves: T25223
Signed-off-by: Simon McVittie <smcv@collabora.com>