runtime: Factor out multiarch_details_get_libdirs
We return the library directories in most-important-first order here,
because that seems like the one that makes most sense (and matches
LD_LIBRARY_PATH). When we go looking for DRI drivers, we actually want
to take the least important first, but iteration over a GPtrArray in
reverse order is easy.

Signed-off-by: Simon McVittie <smcv@collabora.com>