runtime: Do not dlopen layers if it is not needed
Currently the Vulkan layer loader does not support having multiple layer
manifests with the same "name" field.
This can be a problem when you want to support multiple ABIs.

For this reason some layers, like vkBasalt, use a "library_path" with
just the filename of the library. In this way, at execution time, the
correct library will be chosen automatically.

To expand the special tokens LIB, PLATFORM and ORIGIN we dlopened the
layers. But instead of always doing it, now we dlopen layers only if
they actually contain at least one of those tokens.
In this way we can speed up the whole process a little bit and also keep
basenames as is, so that their correct library can still be chosen
automatically at execution time.

Partial addresses #39

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>