pressure-vessel: Make graphics drivers outside $HOME available, if used
Since !173, we have been able to create symbolic links to graphics
drivers anywhere on the filesystem, but that didn't mean the graphics
driver would necessarily be usable, because the directory containing it
would not necessarily be visible in the container. Add these directories
to the list of directories to be "exported".

Manual test (Debian on on Intel hardware, adjust as required for
others):

* Modify /usr/share/vulkan/icd.d/intel_icd.i686.json
  to use /opt/moved-from-usr/lib/i386-linux-gnu/libvulkan_intel.so
* Move the real i386 libvulkan_intel.so to that location
* ./run --verbose -- steam-runtime-system-info 2>&1 | tee container.log

Resolves: #29
Resolves: https://github.com/ValveSoftware/steam-runtime/issues/313
Signed-off-by: Simon McVittie <smcv@collabora.com>