graphics: Use the right filename for Vulkan layers in a sysroot
The API is that when we're inspecting a sysroot, we put the filename
relative to the sysroot in the result object. We did this correctly
for the ICDs, but not for layers. The result was that adding Vulkan
layer support caused a regression for the sandbox-escaping prototype of
launching pressure-vessel from inside Flatpak.

Signed-off-by: Simon McVittie <smcv@collabora.com>