pressure-vessel-wrap: Remove --host-fallback option
We are not going to be able to do this from within Flatpak, and it is
unsuitable for running runtimes other than Steam Runtime 1 'scout',
for which the LD_LIBRARY_PATH runtime mostly already works.

If we need an automatic fallback to running on the host for scout games,
we can do that better by running `pressure-vessel-wrap --test`.

Signed-off-by: Simon McVittie <smcv@collabora.com>