_start-container-in-background: Don't repeatedly start container
If starting the container failed the first time we tried it, there's
no point in trying again repeatedly. Unfortunately, when Steam runs
setup commands, it ignores their exit status, so we can't make it
give up early.

Because the main game runs in a slightly different environment, it's
possible that it might succeed even when setup commands fail, and it's
also possible that launch options are making it more verbose or
debuggable, so try harder to start that one.

Signed-off-by: Simon McVittie <smcv@collabora.com>