runtime: Collect s2tc with correct precedence
Because capsule-capture-libs doesn't overwrite existing symbolic links,
in the unlikely event that s2tc exists in more than one directory in
the search path, we want to look at the highest-precedence first.

The reversed order was correct for the current implementation of
collecting DRI drivers, but not for s2tc.

Signed-off-by: Simon McVittie <smcv@collabora.com>