runtime: Cope with /tmp being a symlink
If /tmp is a symlink in the current execution environment, then when
we ask bwrap to bind-mount /tmp/pressure-vessel-wrap.XXXXXX/overrides
over itself, it will try to create the parent directory /tmp, and fail
because there's a symlink in the way. Canonicalize the path and use that
instead.

Partially addresses
https://github.com/ValveSoftware/steam-runtime/issues/321
(symlinks "above" the home directory have the same problem, but are
harder to fix).

Signed-off-by: Simon McVittie <smcv@collabora.com>