tests: Don't assert that "extra" DRI drivers are in the container
These are drivers for which our best guess is that the native loader
would not actually load them. We want to include them in
steam-runtime-system-info's diagnostic output in case we are wrong, but
we don't want to include them in the container: if our guess is correct,
then the host system doesn't actually load them, and the worst-case
scenario is that they are somehow broken, in which case adding them to
the search path might *break* the container.

Signed-off-by: Simon McVittie <smcv@collabora.com>