runtime: correctly handle ICD for different ABI even without subdirs
Previously we used the empty directory as an indicator that the capture
of the library failed, but this worked only if we always started with an
empty directory. And this is not the case when we are not using a
subdir.

For this reason, even if the capture of a library failed, we didn't set
its "kinds" to "NONEXISTENT", and this could have leaded to errors like
having duplicated Vulkan implicit layer JSON for the same name and
different "library_path", one pointing to the correct ABI and the other
to the wrong one.

Fixes: #45

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>