pressure-vessel: Move environment variables into new PvEnviron object
This keeps track of environment variables' values and their locked
status in one place. We build it up over time, and then serialize it
into final_argv->envp.

In particular, this is a step towards no longer storing anything in the
variable named bwrap except for bubblewrap options, which will be
a necessary piece of refactoring when we move to launching the game
as a Flatpak sub-sandbox, without ever invoking bubblewrap ourselves.

Signed-off-by: Simon McVittie <smcv@collabora.com>