Add support to divert log to a file
With this change it should be easier for users to gather the necessary
logs when filing an issue report.

They can now just launch Steam with "STEAM_LINUX_RUNTIME_LOG=1", and
optionally also "STEAM_LINUX_RUNTIME_VERBOSE=1", and a log file will be
created in their Scout/Soldier install directory.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>