pressure-vessel: Start to use the "info" log level
Now that we have the ability to set the log level to "info", this commit
promotes, from the debug level, all the messages that can be useful to
have, for debugging purposes, but that are not too lousy.

Over time, when we also gain more first hand experience with the issue
reports that users provide, we can tweak the log messages even more.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>