_v2-entry-point: use the relaunch container option by default
The relaunch container option allows us to correctly handle the final
environment variables from the main program, because we don't know them
in advance, when we create the container for the setup steps.

The downside is a potentially increased startup time.

Until we enable this by default it would be very difficult to estimate
the real impact that this option could have on real world systems.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>